Crate integraal

Integraal

Integraal aims to provide generic and efficient tools for numerical integration in the Rust Programming Language.

Quickstart

Multiple standalone examples are provided in the GitHub repository. You can also look at the example provided for the Integraal structure for a very concise overview of the crate’s usage.

Features

As a rule of thumb, a computation method will be gated behind a feature if (a) it requires additional dependencies, or (b) it is not implemented for all integral definitions (e.g. a method that requires a uniform domain).

The following features can be enabled:

boole – enable the Boole computation method.

montecarlo – enable the Monte-Carlo computation method.

romberg – enable the Romberg computation method.

Structs

Integraal

Main integral computation structure

Enums

ComputeMethod

Numerical integration method enum

DomainDescriptor

Domain description enum

FunctionDescriptor

Function description enum

IntegraalError

Integral error

Traits

Scalar

Scalar value trait.